making test coverall pass
