--- .github/workflows/test.yml
@@ -1,41 +1,0 @@
-name: PostgreSQL Test
-
-on: [push]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    services:
-      postgres:
-        image: postgres
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: testdb
-        ports:
-          - 5432:5432
-        options: --health-cmd pg_isready --health-interval 10s --health-timeout 5s --health-retries 5
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Setup Node.js
-        uses: actions/setup-node@v2
-        with:
-          node-version: '19.x'
-      - name: Install dependencies
-        run: npm install
-      - name: Run database migration
-        env:
-          DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          DEV_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          TEST_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          NODE_ENV: development
-        run:  npm run db:migrate
-      - name: Run tests and coverage
-        env:
-          DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          DEV_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          TEST_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
-          CI: true
-          NODE_ENV: development
-        run:  npm run test:coverage && npm run coverage
