+++ .github/workflows/testing.yml
@@ -1,55 +1,53 @@
-name: Test
+name: PostgreSQL Test
 
-on:
-  push:
-    branches: 
-      - develop
-      - main
-    paths-ignore:
-      - '**.md'
-  pull_request:
-    branches: [develop,main]
-    paths-ignore:
-      - '**.md'
-env:
-  NODE_ENV: test
-  TEST_DATABASE_URL: postgres://thunder_user:58hAIQS77klG6uZf58NceFDV57btJN1J@dpg-ce2cksg2i3mkuc8r036g-a.oregon-postgres.render.com/hr_app_test?sslmode=verify-ca
-
+on: [push]
 
 jobs:
   test:
     runs-on: ubuntu-latest
-    steps:
-      -  uses: actions/checkout@v2
-      - name: Build the stack
+    services:
+      postgres:
+        image: postgres
         env:
-          NODE_ENV: test
-          PORT: 3400
-        run: docker-compose -f docker-compose.yaml up -d db
-      -  name: Check folder
-         run: pwd
+          POSTGRES_USER: postgres
+          POSTGRES_PASSWORD: postgres
+          POSTGRES_DB: testdb
+        ports:
+          - 5432:5432
+        options: --health-cmd pg_isready --health-interval 10s --health-timeout 5s --health-retries 5
+    steps:
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Setup Node.js
         uses: actions/setup-node@v2
         with:
           node-version: '19.x'
-
-      # - name: Wait for 1 minutes
-      #   run: sleep 60
-
       - name: Install dependencies
         run: npm install
-
+      - name: Run database migration
+        env:
+          DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
+          DEV_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
+          TEST_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
+          NODE_ENV: development
+        run:  npm run db:migrate
       # - name: Run tests and coverage
       #   env:
-      #     DATABASE_URL: postgres://admin:ptvcNsayeHQO01PXqqQRU754zzPM65BQ@dpg-cfnki6ta499f28c8pvd0-a.oregon-postgres.render.com/codehills_test
-      #     DEV_DATABASE_URL: postgres://admin:ptvcNsayeHQO01PXqqQRU754zzPM65BQ@dpg-cfnki6ta499f28c8pvd0-a.oregon-postgres.render.com/codehills_test
-      #     TEST_DATABASE_URL: postgres://admin:ptvcNsayeHQO01PXqqQRU754zzPM65BQ@dpg-cfnki6ta499f28c8pvd0-a.oregon-postgres.render.com/codehills_test
       #     CI: true
-      #     NODE_ENV: production
+      #     NODE_ENV: test
+      #     HOST: "http://localhost:2023"
+      #     PORT: 2023
+      #     API_VERSION: v1
+      #     FRONTEND_URL: http://localhost:3000
+      #     JWT_SECRET: Vl9fTsn0XUC4SSudWkVp1zwvG8kKPUui
+      #     EXPIRES_IN: 30d
+      #     SESSION_KEY : sadjnsaosada
+      #     CLIENTID : 35ca86da-0c16-461b-ba0d-856abf5c299f
+      #     CLIENTSECRET : MtK8Q~9jcPK321HtjKGiuBIcgL.rmejDvLBoNcQ7
+      #     DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
+      #     DEV_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
+      #     TEST_DATABASE_URL: postgres://postgres:postgres@0.0.0.0:5432/testdb
       #   run:  npm run test:coverage && npm run coverage
-
       # - name: Upload coverage report to Codecov
       #   uses: codecov/codecov-action@v2
       #   with:
@@ -59,67 +57,3 @@ jobs:
       #     path-to-lcov: ./coverage/lcov.info
       # - name: Coveralls Finished
       #   run: env GITHUB_TOKEN=${{ secrets.GITHUB_TOKEN }} yarn coverage --if-present
-
-# name: Test
-
-# on: [push]
-
-# jobs:
-#   test:
-#     runs-on: ubuntu-latest
-
-#     services:
-#       db:
-#         image: postgres
-#         env:
-#           POSTGRES_USER: myuser
-#           POSTGRES_PASSWORD: mypassword
-#           POSTGRES_DB: mydatabase
-#         ports:
-#           - 5432:5432
-
-#     steps:
-#       - name: Checkout code
-#         uses: actions/checkout@v2
-
-#       - name: Install dependencies
-#         run: npm install
-        
-#       # - name: Install dependencies
-#       #   run: |
-#       #     # Create connection string URL
-#       #     connectionString="postgres://${PGUSER}:${PGPASSWORD}@${PGHOST}:${PGPORT}/${PGDATABASE}"
-
-#       #     # Run tests using connection string URL
-#       #     npm test -- --database-url=$connectionString
-
-#       # - name: Run tests
-#       #   env:
-#       #     DATABASE_URL: postgres://myuser:mypassword@localhost:5432/mydatabase
-#       #     DEV_DATABASE_URL: postgres://postgres:example@db/mydb
-#       #     TEST_DATABASE_URL: postgres://myuser:mypassword@localhost:5432/mydatabase
-#       #   run: npm test
-#       - name: Create connection string URL
-#         id: connection-string
-#         run: echo "::set-output name=url::postgres://${{ env.POSTGRES_USER }}:${{ env.POSTGRES_PASSWORD }}@${{ env.DB_HOST }}:${{ job.services.db.ports['5432'] }}/${{ env.POSTGRES_DB }}"
-
-#       - name: Test database connection
-#         env:
-#           DATABASE_URL: ${{ steps.connection-string.outputs.url }}
-#         run: |
-#           psql $DATABASE_URL -c "SELECT 1;
-
-#       - name: Test database connection
-#         env:
-#           DATABASE_URL: ${{ steps.connection-string.outputs.url }}
-#           TEST_DATABASE_URL: ${{ steps.connection-string.outputs.url }}
-#           NODE_ENV: test
-#         run: npm run db:migrate
-
-#       - name: Run tests
-#         env:
-#           DATABASE_URL: ${{ steps.connection-string.outputs.url }}
-#           DEV_DATABASE_URL: postgres://postgres:example@db/mydb
-#           TEST_DATABASE_URL: ${{ steps.connection-string.outputs.url }}
-#           NODE_ENV: test
-#         run: npm run test
