+++ src/restful/controllers/userControllers.js
@@ -1,6 +1,7 @@
 import Response from "./../../system/helpers/Response";
 import DB from "./../../database";
 import UserService from "../../services/userService";
+import { fileUploader } from "../../system/fileUploader";
 const { User } = DB;
 
 export default class UserControllers {
@@ -180,6 +181,64 @@ export default class UserControllers {
     } catch (error) {
       console.log(error);
       return res.status(500).json({ message: "Server error" });
+    }
+  }
+
+  static async update(req, res) {
+    try {
+      const { email } = req.user;
+      const userExist = await UserService.findOneUser({ email });
+      if (!userExist) {
+        return Response.error(res, 404, {
+          message: "The user you are trying to update is not found!",
+        });
+      }
+      let obj = req.body;
+      let avatar = req?.files
+        ? fileUploader(req?.files?.avatar).fileName
+        : userExist?.avatar;
+      obj.avatar = avatar;
+
+      obj.bank = req.body?.bank ? JSON.parse(obj.bank) : userExist?.bank;
+      obj.bank.accountName = req.body?.bank.accountName
+        ? obj.bank.accountName
+        : userExist?.bank.accountName;
+      obj.bank.BankName = req.body?.bank.BankName
+        ? obj.bank.BankName
+        : userExist?.bank.BankName;
+      obj.bank.SwiftCode = req.body?.bank.SwiftCode
+        ? obj.bank.SwiftCode
+        : userExist?.bank.SwiftCode;
+      obj.bank.Currency = req.body?.bank.Currency
+        ? obj.bank.Currency
+        : userExist?.bank.Currency;
+
+      obj.address = req.body?.address
+        ? JSON.parse(obj.address)
+        : userExist?.address;
+      obj.address.country = req.body?.address?.country
+        ? obj.address?.country
+        : userExist?.address?.country;
+      obj.address.city = req.body?.address?.city
+        ? obj.address?.city
+        : userExist?.address?.city;
+      obj.address.street = req.body?.address?.street
+        ? obj.address?.street
+        : userExist?.address?.street;
+
+      await userExist.set(obj);
+      await userExist.save();
+
+      userExist.avatar = `${process.env.HOST}/uploads/${obj?.avatar}`;
+      return Response.success(res, 200, {
+        message: "profile updated successfully",
+        data: userExist,
+      });
+    } catch (error) {
+      return Response.error(res, 500, {
+        message: "server error",
+        error: error.message,
+      });
     }
   }
 }
