+++ src/restful/routes/profile.js
@@ -1,11 +1,12 @@
 import express from "express";
 import AuthController from "../controllers/authControllers";
+import UserControllers from "../controllers/userControllers";
 import protect from "../middlewares";
 
 const router = express.Router();
 
 router.get("/", protect, AuthController.profile);
 
-router.put("/", protect, AuthController.update);
+router.put("/", protect, UserControllers.update);
 
 export default router;
